add readonly, relocate FakeChat to Chat, use english
