+++ src/message/ChainedMessage.test.ts
@@ -1,17 +1,17 @@
-import FakeChat from "../chat/FakeChat";
 import ChainedMessage from "./ChainedMessage";
 import TextMessage from "./TextMessage";
 import FileMessage from "./FileMessage";
 import File from "../file/File";
+import {FakeChat} from "../chat/Chat";
 
 describe("ChainedMessage", () => {
-  it("должен отправить несколько сообщений последовательно через FakeChat", async () => {
+  it("must send multiple messages in sequence", async () => {
     const fakeChat = new FakeChat();
 
     const textMessage = new TextMessage("First message");
     const mockFile = {} as File;
     const fileMessage = new FileMessage(mockFile);
-    const chainedMessage = new ChainedMessage(textMessage, fileMessage);
+    const chainedMessage = new ChainedMessage([textMessage, fileMessage]);
 
     await chainedMessage.sendTo(fakeChat);
