+++ src/message/ChainedMessage.ts
@@ -2,11 +2,9 @@ import Chat from "../chat/Chat";
 import Message from "./Message";
 
 export default class ChainedMessage implements Message {
-  private messages: Message[];
-
-  constructor(...messages: Message[]) {
-    this.messages = messages;
-  }
+  constructor(
+    private readonly messages: Message[]
+  ) {}
 
   async sendTo(context: Chat): Promise<void> {
     for (const message of this.messages) {
