+++ src/message/FileMessage.ts
@@ -3,11 +3,10 @@ import Message from "./Message";
 import File from "../file/File";
 
 export default class FileMessage implements Message {
-  private file: File;
 
-  constructor(file: File) {
-    this.file = file;
-  }
+  constructor(
+    private readonly file: File
+) {}
 
   async sendTo(context: Chat): Promise<void> {
     await context.sendFile(this.file);
