+++ src/message/TextMessage.ts
@@ -2,11 +2,10 @@ import Chat from "../chat/Chat";
 import Message from "./Message";
 
 export default class TextMessage implements Message {
-  private text: string;
 
-  constructor(text: string) {
-    this.text = text;
-  }
+  constructor(
+    private readonly text: string
+) {}
 
   async sendTo(context: Chat): Promise<void> {
     await context.sendText(this.text);
